app/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
flask-mysqldb==0.2.0
mysqlclient==2.1.0
mysql-connector-python==8.0.28